+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 
 - [ui] BModal can be closed with esc key
 - [ui] BSlider replaces BHorizontalList
+- [ui] New request composable